requirements.txt
Django==4.2.9
djangorestframework==3.15.1
django-cors-headers==4.0.0
python-dotenv==1.1.0
Pillow==10.0
djangorestframework-simplejwt===5.3.1
